+++ RecoTracker/ConversionSeedGenerators/plugins/CombinedHitQuadrupletGeneratorForPhotonConversion.h
@@ -14,7 +14,7 @@ class HitQuadrupletGeneratorFromLayerPairForPhotonConversion;
 class SeedingLayerSetsHits;
 namespace edm { class Event; class EventSetup; }
 
-#include "RecoTracker/ConversionSeedGenerators/interface/ConversionRegion.h"
+#include "ConversionRegion.h"
 
 /** \class CombinedHitQuadrupletGeneratorForPhotonConversion
  * Hides set of HitQuadrupletGeneratorFromLayerPairForPhotonConversion generators.
