+++ RecoTracker/ConversionSeedGenerators/plugins/Conv4HitsReco2.cc
@@ -8,9 +8,9 @@
 
 #define Conv4HitsReco2_cxx
 
-//#include "RecoTracker/ConversionSeedGenerators/interface/Conv4HitsReco2.h"
+//#include "Conv4HitsReco2.h"
 //#include "FWCore/MessegeLogger/interface/MessegeLogger.h"
-#include "RecoTracker/ConversionSeedGenerators/interface/Conv4HitsReco2.h"
+#include "Conv4HitsReco2.h"
 #include <time.h>
 
 Conv4HitsReco2::Conv4HitsReco2(math::XYZVector &vPhotVertex, math::XYZVector &h1, math::XYZVector &h2, math::XYZVector &h3, math::XYZVector &h4)
