+++ RecoTracker/ConversionSeedGenerators/plugins/Quad.h
