+++ RecoTracker/ConversionSeedGenerators/plugins/SeedForPhotonConversion1Leg.h
@@ -2,7 +2,7 @@
 #define SeedForPhotonConversion1Leg_H
 
 #include "RecoTracker/TkSeedingLayers/interface/SeedingHitSet.h"
-#include "RecoTracker/ConversionSeedGenerators/interface/PrintRecoObjects.h"
+#include "PrintRecoObjects.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "DataFormats/TrajectorySeed/interface/TrajectorySeedCollection.h"
 #include "FWCore/Framework/interface/EventSetup.h"
