+++ RecoTracker/ConversionSeedGenerators/plugins/TrackChargeSelector.cc
@@ -1,6 +1,6 @@
 #include "FWCore/Framework/interface/MakerMacros.h"
 #include "CommonTools/UtilAlgos/interface/SingleObjectSelector.h"
-#include "RecoTracker/ConversionSeedGenerators/interface/TrackChargeSelector.h"
+#include "TrackChargeSelector.h"
 #include "DataFormats/TrackReco/interface/TrackFwd.h"
 #include "DataFormats/TrackReco/interface/Track.h"
